Allow MCU MBOX SRAM to be staging area for Image Loading and FW Update (#422)
* Allow MCU MBOX SRAM to be staging area for Image Loading and FW Update

Currently Image Loading and FW Update uses an external test sram to pass
firmware images to Caliptra. Since these are not actual HW devices, the
MCU MBOX SRAMs will be used the actual device testing.

Changes include:

1. Creation of a runtime stack to allow direct access to the MCU MBOX
SRAM. This includes the syscall, capsules and driver.

2. Add support for MCU MBOX1 SRAM in the emulator.

3. Add DMA support to MCU MBOX SRAMs. This allows MCU and Caliptra to
DMA from/to the SRAMs.

4. Replace EXTERNAL_TEST_SRAM with MCU_MBOX1_SRAM as staging address for
image loading and firmware update.

* Addressed comments

1. In MCU SRAM driver, defer upcall using Alarm
2. Fix current_app not set in driver
3. Make it explicit why we need to release MCU SRAM lock after reboot

* Update caliptra-sw SHA

- Update Caliptra SHA dependency
- Create a separate alarm instance for MBOX SRAM
- Make sure current_app is cleared only in the upcall

+++ emulator/app/mcu-mbox/src/mcu_mailbox_transport.rs
@@ -1,6 +1,7 @@
 // Licensed under the Apache-2.0 license
 
-use emulator_periph::{McuMailbox0External, MCU_MAILBOX0_SRAM_SIZE};
+use emulator_consts::MCU_MAILBOX0_SRAM_SIZE;
+use emulator_periph::McuMailbox0External;
 use registers_generated::mci::bits::MboxExecute;
 use tock_registers::interfaces::Readable;
 
@@ -15,7 +16,7 @@ impl McuMailboxTransport {
     }
 
     pub fn execute(&self, cmd: u32, payload: &[u8]) -> Result<(), McuMailboxError> {
-        if payload.len() > MCU_MAILBOX0_SRAM_SIZE {
+        if payload.len() > MCU_MAILBOX0_SRAM_SIZE as usize {
             return Err(McuMailboxError::Overflow);
         }
